ops_api/requirements.txt
fastapi==0.116.1
slowapi==0.1.9
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.12
Pillow==10.4.0
httpx==0.27.2
aiofiles>=23.0.0
supabase>=2.0.0
google-auth==2.37.0
